% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Get geo boundary data}
\usage{
get_boundaries(type = "la")
}
\arguments{
\item{type}{The type of boundary to return. la (local authority) by default.}
}
\value{
sf object representing LSOA boundaries
}
\description{
Get geo boundary data
}
\examples{
las = get_boundaries()
# lsoas = get_boundaries(type = "lsoa")
# plot(lsoas$geometry[1:999])
jts_04_emp = get_jts(type = "jts04", purpose = "employment", sheet = 2017)
summary(las_in_jts <- las[[1]] \%in\% jts_04_emp$LA_Code)
las[[3]][!las_in_jts] # Welsh LAs
}
