% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jts.R
\name{get_jts}
\alias{get_jts}
\title{Get JTS data}
\usage{
get_jts(
  type = "jts04",
  purpose = "",
  sheet = "2019",
  table_code = NULL,
  base_url = "https://github.com/ITSLeeds/jts/releases/download/2/"
)
}
\arguments{
\item{type}{What type of JTS data is of interest? One of
"jts01" "jts02" "jts03" "jts04" "jts05" "jts09" "jts10"
jts05, Journey times to key services by lower super output area (JTS05),
provided by default}

\item{purpose}{The purpose or text string matching the table of interest.
Purpose can match 'trip attractors' for which journey time
statistics are required such as
employment
primary
secondary
further
gps
hospitals
food
town
pharmacy.
The text string can also match unique strings to identify tables in each
table type, e.g. setting this to 'rural' when type is set to 'jts01'
will match the table
JTS0102: Average minimum travel time to reach the nearest key services by
mode of travel, rural and urban areas: England.}

\item{sheet}{The sheet within the JTS dataset for which data is required.
Can be a year, with 2014 2015 2016 2017 or 2019 (the default), or a
text string to identify a particular sheet (e.g. "meta" to get metadata tables)}

\item{table_code}{If you already know the JTS table code you're after}

\item{base_url}{Where to get the data from}
}
\description{
Get JTS data
}
\examples{
employment_2017 = get_jts(type = "jts04", purpose = "employment", sheet = 2017)
names(employment_2017)
nrow(employment_2017) # 32844
employment_2017
employment_metadata = get_jts(type = "jts04", purpose = "employment", sheet = "meta")
# View(employment_metadata)
primary_2017 = get_jts(type = "jts04", purpose = "prim", sheet = 2017)
# View(employment_2017) # View results
employment_2017_lsoa = get_jts(type = "jts05", purpose = "employment", sheet = 2017)
# get_jts(sheet = "0") # Error message
gps_2017 = get_jts(purpose = "gp", sheet = 2017)
town = get_jts(purpose = 'town')
# get_jts(type = "jts01", sheet = "by mode") # asks for more info
# get_jts(type = "jts01", sheet = "by mode of travel, local authority", purpose = "")
jts_01_rural_2015 = get_jts(type = "jts01", purpose = "rural", sheet = "2015")
}
