% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jts.R
\name{get_jts_metadata}
\alias{get_jts_metadata}
\title{Get metadata associated with a particular JTS table}
\usage{
get_jts_metadata(
  type = "jts05",
  purpose = "employment",
  sheet = "meta",
  base_url = "https://github.com/ITSLeeds/jts/releases/download/2/"
)
}
\arguments{
\item{type}{What type of JTS data is of interest? One of
"jts01" "jts02" "jts03" "jts04" "jts05" "jts09" "jts10"
jts05, Journey times to key services by lower super output area (JTS05),
provided by default}

\item{purpose}{The purpose or text string matching the table of interest.
Purpose can match 'trip attractors' for which journey time
statistics are required such as
employment
primary
secondary
further
gps
hospitals
food
town
pharmacy.
The text string can also match unique strings to identify tables in each
table type, e.g. setting this to 'rural' when type is set to 'jts01'
will match the table
JTS0102: Average minimum travel time to reach the nearest key services by
mode of travel, rural and urban areas: England.}

\item{sheet}{The sheet within the JTS dataset for which data is required.
Can be a year, with 2014 2015 2016 2017 or 2019 (the default), or a
text string to identify a particular sheet (e.g. "meta" to get metadata tables)}

\item{base_url}{Where to get the data from}
}
\value{
A data frame containing metadata for JTS data
}
\description{
Get metadata associated with a particular JTS table
}
\examples{
get_jts_metadata(type = "jts05", purpose = "employment")
}
