% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jts.R
\docType{data}
\name{jts_params}
\alias{jts_params}
\title{JTS table parameters}
\format{
An object of class \code{list} of length 5.
}
\usage{
jts_params
}
\description{
The complete list of JTS tables represented as named sheets from the ODS files
}
\examples{
jts_params
}
\keyword{datasets}
