% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jts.R
\docType{data}
\name{jts_tables}
\alias{jts_tables}
\title{JTS tables}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 192 rows and 6 columns.
}
\usage{
jts_tables
}
\description{
The complete list of JTS tables represented as named sheets from the ODS files
}
\examples{
jts_tables
}
\keyword{datasets}
