% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jts.R
\name{lookup_jts_table}
\alias{lookup_jts_table}
\title{Find specific jts table and sheet}
\usage{
lookup_jts_table(
  type = "",
  purpose = "",
  sheet = "",
  code = "",
  file_name = ""
)
}
\arguments{
\item{type}{What type of JTS data is of interest? One of
"jts01" "jts02" "jts03" "jts04" "jts05" "jts09" "jts10"
jts05, Journey times to key services by lower super output area (JTS05),
provided by default}

\item{purpose}{The purpose or text string matching the table of interest.
Purpose can match 'trip attractors' for which journey time
statistics are required such as
employment
primary
secondary
further
gps
hospitals
food
town
pharmacy.
The text string can also match unique strings to identify tables in each
table type, e.g. setting this to 'rural' when type is set to 'jts01'
will match the table
JTS0102: Average minimum travel time to reach the nearest key services by
mode of travel, rural and urban areas: England.}

\item{sheet}{The sheet within the JTS dataset for which data is required.
Can be a year, with 2014 2015 2016 2017 or 2019 (the default), or a
text string to identify a particular sheet (e.g. "meta" to get metadata tables)}

\item{code}{If you already know the JTS table code you're after}

\item{file_name}{If you know the unique file name of the JTS table you're after.
This is provided as a character string with two components: the
JTS table code such as jts0102 and the sheet such as the year or metadata
within the JTS spreadsheet. \code{"jts0102-2014"}, for example,
refers to the sheet 2014 in the JTS spreadsheet JTS0102
(Average minimum travel time to reach the nearest key services by mode of
travel, rural and urban areas: England).}
}
\description{
Find specific jts table and sheet
}
\examples{
lookup_jts_table()
lookup_jts_table(purpose = "employment", sheet = 2017)
jdf = lookup_jts_table(purpose = "employment", sheet = 2019)
jdf$table_url
# browseURL(jdf$table_url)
lookup_jts_table(type = "jts01", purpose = "rural")
lookup_jts_table(purpose = "gp", sheet = "meta")
lookup_jts_table(type = "jts05", purpose = "employment", sheet = "meta")
jts_0102_2014 = lookup_jts_table(file_name = "jts0102-2014")
jts_0102_2014$table_title
}
